[FIX] product_configurator_sale: avoid other module inheritance issues
When product configurator is installed, some other modules then
fail to install. It seems that it happens when they are also extending
sales order lines, to add columns to the nested tree, after the "name"
field.

This is probably a subtle bug in the Odoo core, but the pragmatic
solution seems to simplify the Product Configurator sales order form
extensions.